Refactor db type to have many installs
New wrapper DB type. Don't create an empty file on App init, but if we
get null data back (no file yet), return a new DB type with empty array
of installs.

Project listing uses the most recent entry (assuming the flow is
"install" then "view projects"). Realistically the app would have to
know who the user is and use the correct install data, but this is fine
for a demo.