chore: remove name from package.json
There was a malicious package published to npm using the same name as
this package, it has since been taken down but it still exists in Snyk's
vulndb which gives a false positive for anyone scanning this package.

Since this package is not published on npm, nor do we intend to do so,
we can remove the name which associates us to that malicious package.